Create main.html
Worked on this on my own device before deciding to upload it for the sake of delegating tasks.
Has most of the structure and part of the intended functions. A start, but have a list of things to get to before the code can be published and used for regular newsletter publishing.